Updates for tab completions in win32 with no-readline
Tab Completions are fetched from megacmd-server, quoted and partially quoted parameters are supported for commands and autocomplete.
Added `autocomplete` and `codepage` commands, with built in documentation, that work the same way as in `megacli`.
Fixed a possible bug in file version lookup, where local variable `baseNode` was self-referencing in its initialisation.