Apply lessons learned from Synology builds to QNAP.
Prior building of QNAP packages required me to create a directory
hierarchy similar to the develop who built the QNAP toolchain.

This is no longer necessary as we correct the paths present in the
toolchain's .la files.

Another lesson learned from the Synology builds is that we don't need to
run MEGAcmd's autogen/configure scripts. This is safe as we're building
MEGAcmd manually and the configure script isn't generating anything
actually used during build-time.