Fix usage of `MegaApi::startDownload` after new argument added
A new boolean argument (`undelete`) was added to
`MegaApi::startDownload` before the last argument which was a pointer
with a default value of `NULL`.

The usage of this function was not updated in MEGAcmd, but the compiler
didn't complain because the last argument was implicitly being converted
to boolean, and because the pointer has a default value all was good. So
we were effectively calling this function with `undelete=true` and
`listener=NULL` always.