Refactor thread lookups
We've moved the implementation of the thread lookup tables (for stream,
out code, etc.) into a class for the following reasons:
- It'd be easier to unit test and separate to a new file in the future
- Provide strong data-locking guarantees
- Provide a cleaner interface and simpler implementation, that is easier
to use and understand
- Avoid leaking internal details to other methods

The data that the thread lookups contained is now combined into a single
`ThreadData` struct. Thus, we only need one <thread_id, data> map
instead of multiple.

Some of the global functions (such as OUTSTREAM, etc.) have been renamed
to make it clearer that we're accessing/modifying the value for the
CURRENT thread. For example, `setCurrentOutCode` has been renamed to
`setCurrentThreadOutCode`, and so on. These functions are just a
shortcut to the global instance of `ThreadLookupTable`.