Add `FileRotatingLoggedStream` class
At the moment, this class provides a stream implementation that writes
messages to a message buffer, and pops them periodically in a different
thread to write to an output file. It also flushes the file accordingly
when required. Additionally, the message buffer is able to properly
handle Out Of Memory errors.

In the future it will rotate and compress the output file once a certain
size threshold is reached. It will also renew the output file if an
error is detected.