Alternative approach to the underlying memory of the message buffer
This approach to the underlying memory of the message buffer is more
memory-efficient: memory blocks are compact and cache-friendly, they
don't use `malloc`, and pointer chasing is not required to access the
data.

All the data blocks are contiguous in memory, which makes traversing the
message buffer much faster. Also, only the underlying vector needs to
perform heap memory allocations. These allocations are bigger but less
frequent.

The downside to this is that Out-Of-Memory errors are not as precise,
since they happen only on vector allocations instead of on the
construction of each memory block. We can still detect these errors, but
we lose some granularity.