Read current version from MEGACMD_CODE_VERSION instead of file
Since the MegaUpdater executable is also updated with the rest, it
contains the latest MEGAcmd version in the code itself. So we can use
the define to get it, instead of having to access a file (that, as we
saw, might not have been properly written to for whatever reason).

Thus, the `megacmd.version` file is only used to check against the
previous version to broadcast the "MEGAcmd has been updated" message at
the start. This is just an info message (i.e., non-critical) so, if the
file was not written to, there won't be any hammering or other errors.