Read the current version via environment variable or console arg
With this commit, the mega updater checks the current version in this
order:
	1) The `MEGACMD_VERSION_OVERRIDE` environment variable.
	2) The `--version [...]` console argument.
	3) The `MEGACMD_CODE_VERSION` macro.
If one is not provided, it'll fall back to the next one; the default one
being `MEGACMD_CODE_VERSION`, which is hardcoded into the build.

The MEGAcmd server uses (2), whereas the emergency updater (i.e.,
windows/macos scheduled task) uses (3). In any case, (1) can be used to
force an update.