Improve mechanism to wait for output file
When there's an error writing to the output file, we'll wait a bit
before trying to access it again. The time we wait increases
semi-exponentially each time until a max of 1 minute.

Without this mechanism, if there's a persistent issue with the file, we
would hammer the console with constant "can't open output file"
messages.