Load the macOS updater even if unload fails
Currently, the exit code of `unload` is 0 when the updater is not
registered (even though there's an error message). With this commit,
even if the exit code were to be changed to 1 in this case, `load` would
still be called and the updater would be registered correctly.

Also, we added the `-w` option to the loading as this seems to fix an
existing issue.