Fix end quote detection not working when parsing a command
While parsing a command, when we reach a '"' character, we want to find
the closing one to escape whitespaces and other special characters. The
algorithm to find it was not working correctly, which caused the method
to always reach the end of the string.

This would result in, for example: `--arg1="something" --arg2=12` being
parsed together as the string "--arg1=\"something\" --arg2=12". This
means that when we had two arguments, one with quotes and one without,
they would only be parsed correctly if the one with quotes was the last
one.